### 2016-10-27 #99 @whedon
[PRE REVIEW]: rucrdtw: Fast time series subsequence search in R
**Submitting author:** @pboesu (<a href="http://orcid.org/0000-0001-6723-6833">Philipp Boersch-Supan</a>)
**Repository:** <a href="https://github.com/pboesu/rucrdtw" target ="_blank">https://github.com/pboesu/rucrdtw</a>
**Version:** v0.1.0
