### 2016-10-27 #100 @whedon
[REVIEW]: rucrdtw: Fast time series subsequence search in R
**Submitting author:** @pboesu (<a href="http://orcid.org/0000-0001-6723-6833">Philipp Boersch-Supan</a>)
**Repository:** <a href="https://github.com/pboesu/rucrdtw" target ="_blank">https://github.com/pboesu/rucrdtw</a>
**Version:** v0.1.0
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.164995" target="_blank">https://dx.doi.org/10.5281/zenodo.164995</a>
