### 2016-10-27 #99 @arfon
@masalmon - let's continue the review over in #100
