### 2016-10-27 #100 @arfon
@masalmon - please follow the review guidelines (http://joss.theoj.org/about#reviewer_guidelines) and complete the checklist above as you go. Any questions, please don't hesitate to ask!
