### 2016-10-27 #100 @masalmon
Maybe I only accepted for getting issue number 100 ;-)
