### 2016-10-27 #99 @noamross
May I volunteer as a second reviewer for this? I watch the submissions feed for rOpenSci pass-throughs and this one interests me.
