### 2016-10-27 #99 @pboesu
Thanks for you interest @noamross (And of course thanks for agreeing to review @masalmon)

I briefly considered rOpenSci for this, but felt it was in the "standalone statistical tools" category. I'd be more than happy to have a shot in making this compliant with the onboarding criteria, though, if you both think this would fit into the rOpenSci ecosystem!
