### 2016-10-27 #99 @noamross
Hi @pboesu.  I was just telling @arfon that I'm watching this repo to track packages going from RO to JOSS through our joint submittal approach.  You are correct, rucrdtw wouldn't fit in the rOpenSci scope.  I just have done some time series stuff and thought I'd take the chance to explore DTW more deeply by reviewing your package.
