### 2016-10-27 #99 @arfon
> May I volunteer as a second reviewer for this? I watch the submissions feed for rOpenSci pass-throughs and this one interests me.

Absolutely. Please feel free to also contribute your review over in #100 @noamross
