### 2016-10-30 26d6bb1
Change instances of F to FALSE
