### 2016-10-30 pboesu/rucrdtw#2 @noamross
Include tests for failure and coercion
Tests are included and all tests pass, and coverage is high.  I note that where coverage is missing, it is largely because there are not tests that coercion and error handling occur properly ([example](https://codecov.io/gh/pboesu/rucrdtw/src/ad49dddc2eb0c12943a6f057df696b59d9e7e76d/R/ucrdtw.R#L81...82)).  I suggest adding tests to cover these cases, e.g. using `expect_error()`.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/100
