### 2016-10-30 pboesu/rucrdtw#3 @noamross
Improvements to documentation
Documentation in this package is quite sparse, and reflects the idea that users have a detailed understanding of the UCR suite algorithms and API.  My guess is many users will first encounter the UCR suite via this package.   I suggest the following to fill out documentation:
-  [x] A package-level help file (`?rucrdtw` or `?rucrdtw-package`) that gives an overall overview, links to the original UCR web page and reference and vignette.
-  [x] Function-level help files should have examples as well as `@value` fields documenting what the components of the list in function outputs are. This may resemble the "Output Explanation" of the UCR suite README.
-  [x] Argument explanations should be a little more detailed, in particular, greater explanation of the `qlength`, `dtwwindow`arguments. References to the UCR suite paper may be added to keep this from being too long.
-  [x] Other useful information that could be in vignette, README, or package-level
  help file would be performance and functionality comparison to **dtw** or other similar R packages.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/100
