### 2016-10-30 pboesu/rucrdtw#4 @noamross
Exporting a C++ level interface
As this low-level package that may be used as a building block for other work, I suggest exporting the C++ functions so that other packages may import them the `// [[Rcpp::interfaces(r, cpp)]]` attribute marker will generate exportable C++ code.  Other packages will then be able to use the C++ functions without R overhead simply by referring to **rucrdtw** in the `LinkingTo:` field in their DESCRIPTION file.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/100
