### 2016-10-30 #100 @noamross
This is a second review.

---

Thank you for your submission and the opportunity to review this package as a
way of learning more about DTW. This is an excellent, compact package that
provides bindings to a very useful library. I can see its use as a building
block for many applications.

There is one major issue I wish to address here, which is licensing. While the
original UCR source code has re-usage comment, I could not locate a
*license* for the original UCR code which is included in this package.  This
isn't my area of expertise but I want to make sure that re-release under the
Apache license is permitted.

As noted above, there is currently no GitHub release of this package.

I have opened issues on the main repository on [documentation](https://github.com/pboesu/rucrdtw/issues/3)
[testing](https://github.com/pboesu/rucrdtw/issues/2) and [API](https://github.com/pboesu/rucrdtw/issues/4).
