### 2016-10-30 #100 @pboesu
Thanks @noamross for the review and the constructive feedback! I will try to address the specific issues/pull requests in the coming days.

I did want to respond to the issue of licensing, immediately, though, as this is something I was not 100% certain how to handle at the initial submission. As you note, the original source of the wrapped C++ library is copyrighted, but not licensed. I did contact Eamonn Keogh, the senior author of the original code/paper, prior to making my R wrapper repository public on github and he consented to a public release under the Apache license on github and ultimately CRAN.
