### 2016-11-01 6874b56
allowing OSX build failures b/c of travis issues
