### 2016-11-01 pboesu/rucrdtw#3 @maelle
I agree with @noamross :-) I have these further points/comments:
- Another thing I think could be useful (for users who "first encounter the UCR suite via this package") are examples of practical application of the functionalities, ideally one for each of "classification, clustering, and anomaly detection", with or without reference to scientific articles.
- The functions accept different type of data (vector vs. matrix) and format (text files vs. R objects). It'd be great if the "overall overview" of the package-level documentation included a good guide on how to choose a function.
- One particular point that I find misleading/hard to understand is why `qlength` isn't always `length(query)`.
