### 2016-11-01 pboesu/rucrdtw#3 @maelle
Something nice-to-have would be a plot / figure in the README, as marketing strategy.
