### 2016-11-01 #100 @masalmon
This is actually the second review, in theory the first one. ;-)

# Comments:

Thank you for this package!

* I haven't checked the "statement of need" box because I feel there should/could be more details about practical applications.

* There are two examples in the vignette, but they are no "real world" examples, which would really be nice.

* I commented in https://github.com/pboesu/rucrdtw/issues/3 about those two latter points.
