+++ .travis.yml
@@ -26,8 +26,7 @@ matrix:
       - testthat
       - knitr
       - rmarkdown
-      
-allow_failures:
+ allow_failures:
         - os: osx
 
 
