### 2016-11-01 #100 @pboesu
Thank you @masalmon for the review and the PR!

I will try to address all of your comments this week.
