### 2016-11-03 a89f4e0
added package-level documentation
