+++ joss-review-notes.md
@@ -1,0 +1,15 @@
+Joss Review
+===========
+## Issue-2: Tests for failure and coercion
+
+## Issue-3: Documentation
+
+- [x] package-level docs: added in a89f4e003c7cda844bd3fa0dd9cd082c8c068335\
+- [ ] function-level docs - add @value, @examples
+- [ ] more detailed argument explanations
+- [ ] performance and functionality comparison with dtw
+- [ ] Examples of practical applications
+- [ ] guide to file/vector/matrix methods
+- [ ] qlength explanation
+
+## Issue-4: Export C++ level interface
