### 2016-11-03 e3f07ff
removed non-ASCII characters
