### 2016-11-03 140dd3a
exported C++ interface to high-level functions
