+++ src/RcppExports.cpp
@@ -1,44 +1,84 @@
 // Generated by using Rcpp::compileAttributes() -> do not edit by hand
 // Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+#include "../inst/include/rucrdtw.h"
 #include <Rcpp.h>
+#include <string>
+#include <set>
 
 using namespace Rcpp;
 
 // ucrdtw_ff
 Rcpp::List ucrdtw_ff(const char * data, const char * query, int qlength, double dtwwindow);
-RcppExport SEXP rucrdtw_ucrdtw_ff(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEXP, SEXP dtwwindowSEXP) {
+static SEXP rucrdtw_ucrdtw_ff_try(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEXP, SEXP dtwwindowSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const char * >::type data(dataSEXP);
     Rcpp::traits::input_parameter< const char * >::type query(querySEXP);
     Rcpp::traits::input_parameter< int >::type qlength(qlengthSEXP);
     Rcpp::traits::input_parameter< double >::type dtwwindow(dtwwindowSEXP);
     rcpp_result_gen = Rcpp::wrap(ucrdtw_ff(data, query, qlength, dtwwindow));
     return rcpp_result_gen;
-END_RCPP
+END_RCPP_RETURN_ERROR
+}
+RcppExport SEXP rucrdtw_ucrdtw_ff(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEXP, SEXP dtwwindowSEXP) {
+    SEXP rcpp_result_gen;
+    {
+        Rcpp::RNGScope rcpp_rngScope_gen;
+        rcpp_result_gen = PROTECT(rucrdtw_ucrdtw_ff_try(dataSEXP, querySEXP, qlengthSEXP, dtwwindowSEXP));
+    }
+    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
+    if (rcpp_isInterrupt_gen) {
+        UNPROTECT(1);
+        Rf_onintr();
+    }
+    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
+    if (rcpp_isError_gen) {
+        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
+        UNPROTECT(1);
+        Rf_error(CHAR(rcpp_msgSEXP_gen));
+    }
+    UNPROTECT(1);
+    return rcpp_result_gen;
 }
 // ucrdtw_fv
 Rcpp::List ucrdtw_fv(const char * data, Rcpp::NumericVector query, int qlength, double dtwwindow);
-RcppExport SEXP rucrdtw_ucrdtw_fv(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEXP, SEXP dtwwindowSEXP) {
+static SEXP rucrdtw_ucrdtw_fv_try(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEXP, SEXP dtwwindowSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const char * >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::NumericVector >::type query(querySEXP);
[…]
