+++ src/dtw_functions.cpp
@@ -21,6 +21,8 @@
 /***********************************************************************/
 /***********************************************************************/
 
+// [[Rcpp::interfaces(r, cpp)]]
+
 #include <Rcpp.h>
 
 
@@ -34,6 +36,7 @@
 #define min(x,y) ((x)<(y)?(x):(y))
 #define max(x,y) ((x)>(y)?(x):(y))
 #define dist(x,y) ((x-y)*(x-y))
+
 
 #define INF 1e20       //Pseudo Infitinte number for this code
 
