+++ inst/include/rucrdtw_RcppExports.h
@@ -1,0 +1,143 @@
+// Generated by using Rcpp::compileAttributes() -> do not edit by hand
+// Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+#ifndef RCPP_rucrdtw_RCPPEXPORTS_H_GEN_
+#define RCPP_rucrdtw_RCPPEXPORTS_H_GEN_
+
+#include <Rcpp.h>
+
+namespace rucrdtw {
+
+    using namespace Rcpp;
+
+    namespace {
+        void validateSignature(const char* sig) {
+            Rcpp::Function require = Rcpp::Environment::base_env()["require"];
+            require("rucrdtw", Rcpp::Named("quietly") = true);
+            typedef int(*Ptr_validate)(const char*);
+            static Ptr_validate p_validate = (Ptr_validate)
+                R_GetCCallable("rucrdtw", "rucrdtw_RcppExport_validate");
+            if (!p_validate(sig)) {
+                throw Rcpp::function_not_exported(
+                    "C++ function with signature '" + std::string(sig) + "' not found in rucrdtw");
+            }
+        }
+    }
+
+    inline Rcpp::List ucrdtw_ff(const char * data, const char * query, int qlength, double dtwwindow) {
+        typedef SEXP(*Ptr_ucrdtw_ff)(SEXP,SEXP,SEXP,SEXP);
+        static Ptr_ucrdtw_ff p_ucrdtw_ff = NULL;
+        if (p_ucrdtw_ff == NULL) {
+            validateSignature("Rcpp::List(*ucrdtw_ff)(const char *,const char *,int,double)");
+            p_ucrdtw_ff = (Ptr_ucrdtw_ff)R_GetCCallable("rucrdtw", "rucrdtw_ucrdtw_ff");
+        }
+        RObject rcpp_result_gen;
+        {
+            RNGScope RCPP_rngScope_gen;
+            rcpp_result_gen = p_ucrdtw_ff(Rcpp::wrap(data), Rcpp::wrap(query), Rcpp::wrap(qlength), Rcpp::wrap(dtwwindow));
+        }
+        if (rcpp_result_gen.inherits("interrupted-error"))
+            throw Rcpp::internal::InterruptedException();
+        if (rcpp_result_gen.inherits("try-error"))
+            throw Rcpp::exception(as<std::string>(rcpp_result_gen).c_str());
+        return Rcpp::as<Rcpp::List >(rcpp_result_gen);
+    }
+
+    inline Rcpp::List ucrdtw_fv(const char * data, Rcpp::NumericVector query, int qlength, double dtwwindow) {
+        typedef SEXP(*Ptr_ucrdtw_fv)(SEXP,SEXP,SEXP,SEXP);
+        static Ptr_ucrdtw_fv p_ucrdtw_fv = NULL;
+        if (p_ucrdtw_fv == NULL) {
+            validateSignature("Rcpp::List(*ucrdtw_fv)(const char *,Rcpp::NumericVector,int,double)");
+            p_ucrdtw_fv = (Ptr_ucrdtw_fv)R_GetCCallable("rucrdtw", "rucrdtw_ucrdtw_fv");
+        }
+        RObject rcpp_result_gen;
+        {
+            RNGScope RCPP_rngScope_gen;
+            rcpp_result_gen = p_ucrdtw_fv(Rcpp::wrap(data), Rcpp::wrap(query), Rcpp::wrap(qlength), Rcpp::wrap(dtwwindow));
+        }
+        if (rcpp_result_gen.inherits("interrupted-error"))
+            throw Rcpp::internal::InterruptedException();
[…]
