+++ inst/include/rucrdtw.h
@@ -1,0 +1,9 @@
+// Generated by using Rcpp::compileAttributes() -> do not edit by hand
+// Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+#ifndef RCPP_rucrdtw_H_GEN_
+#define RCPP_rucrdtw_H_GEN_
+
+#include "rucrdtw_RcppExports.h"
+
+#endif // RCPP_rucrdtw_H_GEN_
