### 2016-11-03 pboesu/rucrdtw#4 @pboesu
`// [[Rcpp::interfaces(r, cpp)]]` added in 140dd3a240d2978524308b2e8bb7432e56392551

I'm new to Rcpp, so I am not entirely sure how to test whether this worked.
