### 2016-11-03 2127ea6
started adding error tests
