### 2016-11-03 f4afb47
allowed fast_finish on travis
