+++ .travis.yml
@@ -28,6 +28,7 @@ matrix:
       - rmarkdown
  allow_failures:
         - os: osx
+ fast_finish: true
 
 
 notifications:
