### 2016-11-03 e99ba24
added test for missing files, column-wise input, dimension mismatch in vm methods
