+++ tests/testthat/test-errors.R
@@ -8,16 +8,24 @@ data("synthetic_control")
 #cut off some columns
 wrong_ref <- synthetic_control[,1:55]
 
-test_that("ucrdtw_vm error message is displayed on unsuitable input", {
+test_that("error message is displayed on unsuitable input", {
   #input not a numeric matrix
   expect_error(ucrdtw_vm(dataf, firstf, 60, 0.05), "data must be a numeric matrix")
-  #input dimensions wrong
+  expect_error(ucred_vm(dataf, firstf, 60), "data must be a numeric matrix")
+  expect_error(ucrdtw_ff("missing.file", firstf, 60, 0.05), "ERROR : File not Found!!!")
+  expect_error(ucred_ff("missing.file", firstf, 60), "ERROR : File not Found!!!")
 })
 
-test_that("vm method works", {
+test_that("dtw vm method errors on dimension mismatch", {
     query <- synthetic_control[123,]
     #microbenchmark::microbenchmark(
     expect_error(ucrdtw_vm(wrong_ref, query, length(query), 0.05, byrow = TRUE), "data series length must match query length")
     expect_error(ucrdtw_vm(wrong_ref, query, length(query), 0.05, byrow = FALSE), "data series length must match query length")
+})
 
+test_that("dtw ed method works errors on dimension mismatch", {
+  query <- synthetic_control[123,]
+  #microbenchmark::microbenchmark(
+  expect_error(ucred_vm(wrong_ref, query, length(query), byrow = TRUE), "data series length must match query length")
+  expect_error(ucred_vm(wrong_ref, query, length(query), byrow = FALSE), "data series length must match query length")
 })
