### 2016-11-03 da0f6e1
tests for missing query files
