### 2016-11-03 8d00a9c
more tests for missing data files
