### 2016-11-03 42792aa
updated review notes
