+++ joss-review-notes.md
@@ -1,10 +1,11 @@
 Joss Review
 ===========
 ## Issue-2: Tests for failure and coercion
+- [x] added tests for missing input files, wrong input types and dimensions, and columnwise data input for vm methods as of 8d00a9c1083bb1652353c4267a3269684680057c
 
 ## Issue-3: Documentation
 
-- [x] package-level docs: added in a89f4e003c7cda844bd3fa0dd9cd082c8c068335\
+- [x] package-level docs: added in a89f4e003c7cda844bd3fa0dd9cd082c8c068335
 - [ ] function-level docs - add @value, @examples
 - [ ] more detailed argument explanations
 - [ ] performance and functionality comparison with dtw
