### 2016-11-03 pboesu/rucrdtw#2 @pboesu
added tests for missing input files, wrong input types and dimensions, and columnwise data input for vm methods
(Commits up to 8d00a9c1083bb1652353c4267a3269684680057c)
