+++ .travis.yml
@@ -26,6 +26,7 @@ matrix:
       - testthat
       - knitr
       - rmarkdown
+      - dtw
  allow_failures:
         - os: osx
  fast_finish: true
