### 2016-11-04 ca31b2e
rewrote paper to clarify general applications of DTW vs specific application of rucrdtw
