+++ vignettes/rucrdtw.bib
@@ -42,3 +42,14 @@
     year = {1999},
     pages = {27--29}
 }
+
+@article{giorgino2009computing,
+  title={Computing and visualizing dynamic time warping alignments in {R}: the dtw package},
+  author={Giorgino, Toni},
+  journal={Journal of Statistical Software},
+  volume={31},
+  number={7},
+  pages={1--24},
+  year={2009},
+  doi = {10.18637/jss.v031.i07}
+}
