### 2016-11-05 afb979d
updated documentation for dtw_ff method
