### 2016-11-05 cd612f6
added table of methods and use cases
