### 2016-11-05 ef8f1dd
added a draft reviewer response for the github issues
