+++ reviewer-response.md
@@ -1,0 +1,17 @@
+Reviewer Response to docs issue
+
+> Another thing I think could be useful (for users who "first encounter the UCR suite via this package") are examples of practical application of the functionalities, ideally one for each of "classification, clustering, and anomaly detection", with or without reference to scientific articles.
+
+rucrdtw is really only aimed at using DTW for similarity searches. I have rewritten the paper.md to highlight this, and added a reference and link to the `dtw` package. (commit ca31b2e)
+
+> The functions accept different type of data (vector vs. matrix) and format (text files vs. R objects). It'd be great if the "overall overview" of the package-level documentation included a good guide on how to choose a function.
+
+I have added a table with an overview of the different methods and input types into `?rucrdtw-package` (commit cd612f6)
+
+> One particular point that I find misleading/hard to understand is why qlength isn't always length(query).
+
+This is a design feature I retained from the original UCR Suite code, without much thought to it. I am guessing the original underlying idea was to leverage user knowledge to avoid reading the query file twice. However, I see now, how this may be confusing. 
+I decided to remove it as a user input for the fv, vv, and mv methods, given that passing in the query as an R vector enables the user to decide on the query length before the function call. I did, however leave it as an argument in the ff method to avoid reading the query file twice, and to keep the API of that function identical to the command line version of the UCR Suite.
+
+
+
