### 2016-11-05 02008f7
removed qlength argument from vector/matrix methods; adapted docs, tests; expanded docs, examples for dtw methods
