+++ tests/testthat/test-errors.R
@@ -12,26 +12,26 @@ wrong_ref <- synthetic_control[,1:55]
 
 test_that("error message is displayed on unsuitable input", {
   #input not a numeric matrix
-  expect_error(ucrdtw_vm(dataf, firstf, 60, 0.05), "data must be a numeric matrix")
-  expect_error(ucred_vm(dataf, firstf, 60), "data must be a numeric matrix")
+  expect_error(ucrdtw_mv(dataf, firstf, 0.05), "data must be a numeric matrix")
+  expect_error(ucred_mv(dataf, firstf), "data must be a numeric matrix")
   expect_error(ucrdtw_ff("missing.file", firstf, 60, 0.05), "ERROR : File not Found!!!")
   expect_error(ucrdtw_ff(dataf, "missing.file", 60, 0.05), "ERROR : File not Found!!!")
   expect_error(ucred_ff("missing.file", firstf, 60), "ERROR : File not Found!!!")
   expect_error(ucred_ff(dataf, "missing.file", 60), "ERROR : File not Found!!!")
-  expect_error(ucrdtw_fv("missing.file", synthetic_control[1,], qlength, 0.05), "ERROR : File not Found!!!")
-  expect_error(ucred_fv("missing.file", synthetic_control[1,], qlength), "ERROR : File not Found!!!")
+  expect_error(ucrdtw_fv("missing.file", synthetic_control[1,], 0.05), "ERROR : File not Found!!!")
+  expect_error(ucred_fv("missing.file", synthetic_control[1,]), "ERROR : File not Found!!!")
 })
 
 test_that("dtw vm method errors on dimension mismatch", {
     query <- synthetic_control[123,]
     #microbenchmark::microbenchmark(
-    expect_error(ucrdtw_vm(wrong_ref, query, length(query), 0.05, byrow = TRUE), "data series length must match query length")
-    expect_error(ucrdtw_vm(wrong_ref, query, length(query), 0.05, byrow = FALSE), "data series length must match query length")
+    expect_error(ucrdtw_mv(wrong_ref, query, 0.05, byrow = TRUE), "data series length must match query length")
+    expect_error(ucrdtw_mv(wrong_ref, query, 0.05, byrow = FALSE), "data series length must match query length")
 })
 
 test_that("dtw ed method works errors on dimension mismatch", {
   query <- synthetic_control[123,]
   #microbenchmark::microbenchmark(
-  expect_error(ucred_vm(wrong_ref, query, length(query), byrow = TRUE), "data series length must match query length")
-  expect_error(ucred_vm(wrong_ref, query, length(query), byrow = FALSE), "data series length must match query length")
+  expect_error(ucred_mv(wrong_ref, query, byrow = TRUE), "data series length must match query length")
+  expect_error(ucred_mv(wrong_ref, query, byrow = FALSE), "data series length must match query length")
 })
