+++ src/dtw_functions.cpp
@@ -399,7 +399,6 @@ void error(int id)
 //' \itemize{
 //'   \item \strong{location:} The starting location of the nearest neighbor of the given query, of size \code{qlength}, in the data. Note that location starts from 1.
 //'   \item \strong{distance:} The DTW distance between the nearest neighbor and the query.
-//'   \item \strong{exec_time:} Execution time of the query.
 //'   \item \strong{prunedKim:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
 //'   \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
 //'   \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
@@ -436,7 +435,7 @@ Rcpp::List ucrdtw_ff(const char * data , const char * query, int qlength, double
   double ex , ex2 , mean, std;
   int m=-1, r=-1;
   rcpp_long_long_type loc = 0;
-  double t1,t2;
+  //double t1,t2;
   int kim = 0,keogh = 0, keogh2 = 0;
   double dist=0, lb_kim=0, lb_k=0, lb_k2=0;
   double *buffer, *u_buff, *l_buff;
@@ -469,7 +468,7 @@ Rcpp::List ucrdtw_ff(const char * data , const char * query, int qlength, double
     error(2);
 
   /// start the clock
-  t1 = clock();
+  //t1 = clock();
 
 
   /// malloc everything here
@@ -749,13 +748,13 @@ Rcpp::List ucrdtw_ff(const char * data , const char * query, int qlength, double
   free(l_buff);
   free(u_buff);
 
-  t2 = clock();
+  //t2 = clock();
   //Rprintf("\n");
 
   /// Note that loc and i are rcpp_long_long_type.
   Rcpp::List out = Rcpp::List::create(Rcpp::Named("location") = loc + 1, //convert to R's 1-based indexing
                                       Rcpp::Named("distance") = sqrt(bsf),
-                                      Rcpp::Named("exec_time")=(t2-t1)/CLOCKS_PER_SEC,
+                                      //Rcpp::Named("exec_time")=(t2-t1)/CLOCKS_PER_SEC,
                                       Rcpp::Named("prunedKim")=((double) kim / i)*100,
                                       Rcpp::Named("prunedKeogh")=((double) keogh / i)*100,
                                       Rcpp::Named("prunedKeogh2")=((double) keogh2 / i)*100,
@@ -770,15 +769,35 @@ Rcpp::List ucrdtw_ff(const char * data , const char * query, int qlength, double
 //' This implementation of the UCR Suite command line utility, takes a data file as input and an R numeric vector for the query
 //'
 //' @name ucrdtw_fv
-//' @param data char path to data file
-//' @param query numeric vector containing the query
-//' @param qlength int length of query (n data points)
-//' @param dtwwindow double warping window size (as a proportion of query length)
+//' @param data character; path to data file
+//' @param query numeric vector containing the query. The query length is set to the length of this object.
+//' @param dtwwindow double; Size of the warping window size (as a proportion of query length). The DTW calculation in `rucrdtw` uses a symmetric Sakoe-Chiba band. See Giorgino (2009) for a general coverage of warping window constraints.
+//' @return a ucrdtw object. A list with the following elements
+//' \itemize{
+//'   \item \strong{location:} The starting location of the nearest neighbor of the given query, of size \code{qlength}, in the data. Note that location starts from 1.
[…]
