+++ src/RcppExports.cpp
@@ -42,23 +42,22 @@ RcppExport SEXP rucrdtw_ucrdtw_ff(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEX
     return rcpp_result_gen;
 }
 // ucrdtw_fv
-Rcpp::List ucrdtw_fv(const char * data, Rcpp::NumericVector query, int qlength, double dtwwindow);
-static SEXP rucrdtw_ucrdtw_fv_try(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEXP, SEXP dtwwindowSEXP) {
+Rcpp::List ucrdtw_fv(const char * data, Rcpp::NumericVector query, double dtwwindow);
+static SEXP rucrdtw_ucrdtw_fv_try(SEXP dataSEXP, SEXP querySEXP, SEXP dtwwindowSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::traits::input_parameter< const char * >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::NumericVector >::type query(querySEXP);
-    Rcpp::traits::input_parameter< int >::type qlength(qlengthSEXP);
     Rcpp::traits::input_parameter< double >::type dtwwindow(dtwwindowSEXP);
-    rcpp_result_gen = Rcpp::wrap(ucrdtw_fv(data, query, qlength, dtwwindow));
+    rcpp_result_gen = Rcpp::wrap(ucrdtw_fv(data, query, dtwwindow));
     return rcpp_result_gen;
 END_RCPP_RETURN_ERROR
 }
-RcppExport SEXP rucrdtw_ucrdtw_fv(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEXP, SEXP dtwwindowSEXP) {
+RcppExport SEXP rucrdtw_ucrdtw_fv(SEXP dataSEXP, SEXP querySEXP, SEXP dtwwindowSEXP) {
     SEXP rcpp_result_gen;
     {
         Rcpp::RNGScope rcpp_rngScope_gen;
-        rcpp_result_gen = PROTECT(rucrdtw_ucrdtw_fv_try(dataSEXP, querySEXP, qlengthSEXP, dtwwindowSEXP));
+        rcpp_result_gen = PROTECT(rucrdtw_ucrdtw_fv_try(dataSEXP, querySEXP, dtwwindowSEXP));
     }
     Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
     if (rcpp_isInterrupt_gen) {
@@ -75,25 +74,24 @@ RcppExport SEXP rucrdtw_ucrdtw_fv(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEX
     return rcpp_result_gen;
 }
 // ucrdtw_vv
-Rcpp::List ucrdtw_vv(Rcpp::NumericVector data, Rcpp::NumericVector query, int qlength, double dtwwindow, int epoch, bool skip);
-static SEXP rucrdtw_ucrdtw_vv_try(SEXP dataSEXP, SEXP querySEXP, SEXP qlengthSEXP, SEXP dtwwindowSEXP, SEXP epochSEXP, SEXP skipSEXP) {
+Rcpp::List ucrdtw_vv(Rcpp::NumericVector data, Rcpp::NumericVector query, double dtwwindow, int epoch, bool skip);
+static SEXP rucrdtw_ucrdtw_vv_try(SEXP dataSEXP, SEXP querySEXP, SEXP dtwwindowSEXP, SEXP epochSEXP, SEXP skipSEXP) {
 BEGIN_RCPP
[…]

[…]
