### 2016-11-05 7ba2af2
adapted and rebuilt vignette
