### 2016-11-05 45d9e94
minor fixes to examples
