### 2016-11-05 1effe77
disallowed OSX build failures
