+++ .travis.yml
@@ -27,8 +27,6 @@ matrix:
       - knitr
       - rmarkdown
       - dtw
- allow_failures:
-        - os: osx
  fast_finish: true
 
 
