### 2016-11-05 fc6043c
tidied up docs
