+++ R/ucrdtw.R
@@ -36,11 +36,11 @@ summary.ucred <- function(object, ...){
 #'
 #' @name ucrdtw_mv
 #' @param data numeric matrix containing data
-#' @param query numeric vector containing the query
-#' @param dtwwindow double warping window size (as a proportion of query length)
+#' @param query numeric vector containing the query. This determines the query length.
+#' @param dtwwindow double; Size of the warping window size (as a proportion of query length). The DTW calculation in `rucrdtw` uses a symmetric Sakoe-Chiba band. See Giorgino (2009) for a general coverage of warping window constraints.
 #' @param epoch int defaults to 1e5, should be \eqn{\le} 1e6. This is the size of the data chunk that is processed at once. All cumulative values in the algorithm will be restarted after \code{epoch} iterations to reduce floating point errors in these values.
-#' @param skip bool defaults to TRUE. If TRUE bound calculations and if necessary, distance calculations, are only performed on non-overlapping segments of the data (i.e. multiples of \code{qlength}). This is useful if \code{data} is a set of multiple reference time series, each of length \code{qlength}. The location returned when skipping is the index of the subsequence.
-#' @param byrow logical If TRUE rows represent time-series, columns time-points
+#' @param skip boolean; defaults to TRUE. If TRUE bound calculations and if necessary, distance calculations, are only performed on non-overlapping segments of the data (i.e. multiples of \code{length(query)}). This is useful if \code{data} is a set of multiple reference time series, each of length \code{length(query)}. The location returned when skipping is the index of the subsequence.
+#' @param byrow logical; If TRUE rows in \code{data} represent time-series, columns time-points
 #' @return a ucrdtw object. A list containing the following elements
 #' \itemize{
 #'   \item \strong{location:} The row or column number of the nearest neighbor of the given query in the data set.
@@ -50,11 +50,13 @@ summary.ucred <- function(object, ...){
 #'   \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
 #'   \item \strong{dtwCalc:} Percentage of subsequences for which the full DTW distance was calculated.
 #' }
+#' @references Giorgino, Toni (2009). Computing and Visualizing Dynamic Time Warping Alignments in R: The dtw Package. Journal of Statistical Software, 31(7), 1-24, doi:\href{http://dx.doi.org/10.18637/jss.v031.i07}{10.18637/jss.v031.i07}.
 #' @examples
 #' #load example data
 #' data("synthetic_control")
 #' query <- synthetic_control[5,]
-# ucrdtw_mv(synthetic_control, query, 0.05, byrow = TRUE)
+#' #run query
+#' ucrdtw_mv(synthetic_control, query, 0.05, byrow = TRUE)
 #' @export
 ucrdtw_mv <- function(data, query, dtwwindow, epoch = 1e5, skip=TRUE, byrow=FALSE){
   if (class(data)!="matrix" && class(data[1])!="numeric") stop("data must be a numeric matrix")
@@ -72,15 +74,27 @@ ucrdtw_mv <- function(data, query, dtwwindow, epoch = 1e5, skip=TRUE, byrow=FALS
 #' UCR ED Algorithm matrix-vector method
 #'
 #' This implementation of the UCR Suite command line utility, takes an R numeric matrix as data input and an R numeric vector for the query.
-#' The default behaviour differs from the other metjods, in that it does not perform a sliding window search for a match.
+#' The default behaviour differs from the other methods, in that it does not perform a sliding window search for a match.
 #' Instead it is designed to find a best match for a query in a reference set of time-series of the same length as the query.
 #' This is useful, for example, when comparing a time-series of undetermined class to a labelled reference set of classified time-series.
 #'
 #' @name ucred_mv
[…]
