### 2016-11-05 1a0bd1c
added picture to readme
