### 2016-11-05 8618b05
added benchmark to vignette
