### 2016-11-05 c7cf616
wider readme figure
