+++ man/ucrdtw_mv.Rd
@@ -41,6 +41,8 @@ This is useful, for example, when comparing a time-series of undetermined class 
 #load example data
 data("synthetic_control")
 query <- synthetic_control[5,]
+#run query
+ucrdtw_mv(synthetic_control, query, 0.05, byrow = TRUE)
 }
 \references{
 Giorgino, Toni (2009). Computing and Visualizing Dynamic Time Warping Alignments in R: The dtw Package. Journal of Statistical Software, 31(7), 1-24, doi:\href{http://dx.doi.org/10.18637/jss.v031.i07}{10.18637/jss.v031.i07}.
