### 2016-11-05 85b2515
updated review notes
