### 2016-11-05 c77acf6
added figure to paper
