+++ DESCRIPTION
@@ -25,5 +25,6 @@ RoxygenNote: 5.0.1
 Suggests: testthat,
     knitr,
     rmarkdown,
-    dtw
+    dtw,
+    microbenchmark
 VignetteBuilder: knitr
