### 2016-11-05 #100 @pboesu
Hi @masalmon  and @noamross ,
Thank you again for your constructive reviews! I have now completed my revision of the paper and code.

I believe I have addressed all suggestions, except for adding "real world" examples.  A detailed response about the examples, and the other changes to the documentation is [here](https://github.com/pboesu/rucrdtw/issues/3#issuecomment-258646019).

I have also responded to the other two issues about a [C++ level interface](https://github.com/pboesu/rucrdtw/issues/4#issuecomment-258286311) and additional [unit tests](https://github.com/pboesu/rucrdtw/issues/2#issuecomment-258307597).

If you are happy with those changes please let me know. I will then tag the release on github to create an archived version on zenodo, and also submit to CRAN.
