### 2016-11-05 ba921b0
more programmatic approach to finding a suitable example series
