+++ data-raw/ecg_data.R
@@ -13,15 +13,31 @@ rdann -r mitdb/119 -a atr -v >119.ann
 rdann -r mitdb/121 -a atr -v >121.ann
 rdann -r mitdb/230 -a atr -v >230.ann
 
+download_ecg <- function(ecg_id){
+  system(paste("rdsamp -r mitdb/", ecg_id, " -p -v > data-raw/", ecg_id, ".txt", sep=""))
+  system(paste("rdann -r mitdb/", ecg_id, " -a atr -v > data-raw/", ecg_id, ".ann", sep=""))
+}
 
+explore_ecg <- function(ecg_id, start_sec, duration_sec){
+  ecg <- read.table(paste("data-raw/", ecg_id,".txt", sep=""), header =FALSE, skip=2)
+  startrow <- which(ecg[,1]==start_sec)
+  endrow <- which(ecg[,1]==start_sec+duration_sec)
+  range <- startrow:endrow
+  plot(ecg[range,1], ecg[range,2], type="l")
+  lines(ecg[range,1], ecg[range,3], col='blue')
+}
 
-ecg1 <- read.table("data-raw/119.txt", header =FALSE, skip=2)
+download_ecg(200)
+explore_ecg(200, 13*60+45, 6)
+
+ecg1 <- read.table("data-raw/106.txt", header =FALSE, skip=2)
 ecg1_ann <- read.table("data-raw/119.ann", header=FALSE, skip=1,fill=TRUE, col.names = c("Time","Sample","Type","Sub","Chan","Num","Aux"))
 ecg1_ann[ecg1_ann$Type=="V",]$Sample
 
 plot(ecg1[,1], ecg1[,2], type="l")
-range <- 1:2000
+range <- (7*60):(7*60+2000)
 plot(ecg1[range,1], ecg1[range,2], type="l")
+lines(ecg1[range,1], ecg1[range,3], col='blue')
 range_pvc <- (503+30):(503+130)
 lines(ecg1[range_pvc,1], ecg1[range_pvc,2], type="l",col="red")
 
