### 2016-11-06 f67e049
added cran comments
