+++ cran-comments.md
@@ -1,0 +1,30 @@
+## Release summary
+
+This is the first attempted CRAN release of rucrdtw, and my second package submitted to CRAN.
+
+## Test environments
+* OS X 10.11.6 (on travis-ci) R 3.3.2 and R-devel (2016-10-26 r71594)
+* Ubuntu 14.04.4 LTS (on travis-ci) R 3.3.2
+* win-builder (devel and release)
+
+## R CMD check results
+There were no ERRORs or WARNINGs.
+
+There was 1 NOTE on win-builder:
+
+* New submission
+
+  Possibly mis-spelled words in DESCRIPTION:
+  Rakthanmanon (15:68)
+  SIGKDD (15:93)
+  Thanawin (15:59)
+  UCR (3:27, 15:48)
+  al (15:84)
+  et (15:81)
+  subsequence (15:126)
+  ultrafast (15:116)
+
+
+
+* The note was for the package being a first-time submission.
+  Furthermore eight words were highlighted as possibly mis-spelled. The first six result from an in text citation to the C++ library that this submission provides R bindings for, the second two stem from the terminology that is used by the original authors of the C++ library.
