### 2016-11-06 pboesu/rucrdtw#3 @maelle
Nice work @pboesu !
- The package level documentation level is great! The table will be useful I think.
- Your modification regarding `qlength` makes sense.
- The explanation of `prunedKeogh` and `prunedKeogh2` in `ucrdtw_mv` (and other functions I think) is the same as the one for `prunedKim`.
- I understand your arguments regarding real-world example. I guess that at some point you'll add a reference to your work when it's published.
