### 2016-11-06 #100 @masalmon
Hi @pboesu as discussed in the issue in your repo I'm happy with the changes, nice work!
