### 2016-11-06 512fb1f
corrected pruning criteria in docs
