+++ R/RcppExports.R
@@ -15,8 +15,8 @@
 #'   \item \strong{location:} The starting location of the nearest neighbor of the given query, of size \code{length(query)}, in the data. Note that location starts from 1.
 #'   \item \strong{distance:} The DTW distance between the nearest neighbor and the query.
 #'   \item \strong{prunedKim:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
-#'   \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
-#'   \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
+#'   \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Keogh-EQ criterion.
+#'   \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Keogh-EC criterion.
 #'   \item \strong{dtwCalc:} Percentage of subsequences for which the full DTW distance was calculated.
 #' }
 #' For an explanation of the pruning criteria see Rakthanmanon et al. (2012).
@@ -51,8 +51,8 @@ ucrdtw_ff <- function(data, query, qlength, dtwwindow) {
 #'   \item \strong{location:} The starting location of the nearest neighbor of the given query, of size \code{length(query)}, in the data. Note that location starts from 1.
 #'   \item \strong{distance:} The DTW distance between the nearest neighbor and the query.
 #'   \item \strong{prunedKim:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
-#'   \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
-#'   \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
+#'   \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Keogh-EQ criterion.
+#'   \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Keogh-EC criterion.
 #'   \item \strong{dtwCalc:} Percentage of subsequences for which the full DTW distance was calculated.
 #' }
 #' For an explanation of the pruning criteria see Rakthanmanon et al. (2012).
@@ -89,8 +89,8 @@ ucrdtw_fv <- function(data, query, dtwwindow) {
 #'   \item \strong{location:} The starting location of the nearest neighbor of the given query, of size \code{length(query)}, in the data. Note that location starts from 1.
 #'   \item \strong{distance:} The DTW distance between the nearest neighbor and the query.
 #'   \item \strong{prunedKim:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
-#'   \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
-#'   \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
+#'   \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Keogh-EQ criterion.
+#'   \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Keogh-EC criterion.
 #'   \item \strong{dtwCalc:} Percentage of subsequences for which the full DTW distance was calculated.
 #' }
 #' For an explanation of the pruning criteria see Rakthanmanon et al. (2012).
