+++ man/ucrdtw_vv.Rd
@@ -23,8 +23,8 @@ a ucrdtw object. A list with the following elements
   \item \strong{location:} The starting location of the nearest neighbor of the given query, of size \code{length(query)}, in the data. Note that location starts from 1.
   \item \strong{distance:} The DTW distance between the nearest neighbor and the query.
   \item \strong{prunedKim:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
-  \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
-  \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Kim criterion.
+  \item \strong{prunedKeogh:} Percentage of subsequences that were pruned based on the LB-Keogh-EQ criterion.
+  \item \strong{prunedKeogh2:} Percentage of subsequences that were pruned based on the LB-Keogh-EC criterion.
   \item \strong{dtwCalc:} Percentage of subsequences for which the full DTW distance was calculated.
 }
 For an explanation of the pruning criteria see Rakthanmanon et al. (2012).

[…]
