### 2016-11-06 pboesu/rucrdtw#3 @pboesu
Thank you for the quick response!

> The explanation of prunedKeogh and prunedKeogh2 in ucrdtw_mv (and other functions I think) is the same as the one for prunedKim.

and thanks for catching the copy & paste error! I've fixed the documentation in 512fb1fd
