### 2016-11-06 pboesu/rucrdtw#4 @noamross
Looks good.  The important thing is the new files generated in `inst/include`, which can now be included as header files in any other package.
