### 2016-11-06 #100 @noamross
I've examined all of the changes since I first reviewed this package and @pboesu has satisfied all my concerns.  I approve!
