+++ vignettes/using_rucrdtw.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Using rucrdtw"
 author: "Philipp Boersch-Supan"
-date: "`r Sys.Date()`"
+date: "2016-11-06"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Using rucrdtw}
