+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: rucrdtw
 Type: Package
 Title: R Bindings for the UCR Suite
 Version: 0.1.0
-Date: 2016-10-23
+Date: 2016-11-06
 Authors@R: c(person("Philipp", "Boersch-Supan", email = "pboesu@gmail.com", role = c("aut", "cre")),
              person("Thanawin", "Rakthanmanon", role = "aut"),
              person("Bilson", "Campana", role = "aut"),
