### 2016-11-07 #100 @pboesu
Brilliant! Thanks again, @noamross and @masalmon for the constructive review!

I have tagged the release and archived it on zenodo. The archive doi is [10.5281/zenodo.164995](https://doi.org/10.5281/zenodo.164995).

I also have submitted the package to CRAN.
