### 2016-11-07 ee41d71
Added DOI to description for CRAN
