+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rucrdtw
 Type: Package
 Title: R Bindings for the UCR Suite
-Version: 0.1.0
+Version: 0.1.1
 Date: 2016-11-06
 Authors@R: c(person("Philipp", "Boersch-Supan", email = "pboesu@gmail.com", role = c("aut", "cre")),
              person("Thanawin", "Rakthanmanon", role = "aut"),
@@ -12,7 +12,7 @@ Authors@R: c(person("Philipp", "Boersch-Supan", email = "pboesu@gmail.com", role
              )
 BugReports: https://github.com/pboesu/rucrdtw/issues
 URL: https://github.com/pboesu/rucrdtw
-Description: R bindings for functions from the UCR Suite (Thanawin Rakthanmanon et al. 2012 SIGKDD), which enables ultrafast subsequence
+Description: R bindings for functions from the UCR Suite (Thanawin Rakthanmanon et al. 2012 SIGKDD, DOI:10.1145/2339530.2339576), which enables ultrafast subsequence
       search for a best match under Dynamic Time Warping and Euclidean Distance.
 License: Apache License
 LazyData: TRUE
