### 2016-11-07 9982f40
formatted in text citation for CRAN
