### 2016-11-07 c710435
updated CRAN notes
