### 2016-11-07 341eec0
updated cran-comments for 0.1.1
