### 2016-11-07 #100 @arfon
@noamross and @masalmon many thanks for the reviews here ✨

@pboesu - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00100 🎉 🚀 💥
